Fetch ontologies before pydantic validation (#201)
Dojo is currently not publishing to Causemos in `dev` because the
Pydantic schema is not validating. If this patch doesn't work then some
possible solutions include (A) fixing the pydantic schema the
causmosified model, (B) removing the use of pydantic for causemos
publishing entirely as a quick and dirty patch or (C) investigating if
the `get_ontologies` function is corrupting the payload. NOTE: THIS
PATCH HAS NOT BEEN TESTED